requirements.txt.txt
appdirs==1.4.3
asgiref==3.2.7
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cssselect==1.1.0
defusedxml==0.6.0
Django==3.0.4
django-crispy-forms==1.9.0
django-guardian==2.2.0
fake-useragent==0.1.11
gunicorn==20.0.4
idna==2.9
lxml==4.5.0
oauthlib==3.1.0
parse==1.15.0
Pillow==7.1.0
pycparser==2.20
pyee==7.0.1
PyJWT==1.7.1
pyppeteer==0.0.25
pyquery==1.4.1
python-dotenv==0.13.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-html==0.10.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0
sqlparse==0.3.1
tqdm==4.45.0
urllib3==1.25.8
w3lib==1.21.0
websockets==8.1
